moved Gatherer to separate file
